requirements/production.txt
django==2.0.4
pymysql
django-environ